simplify graphql requests - I don't think we need separate types for query/mutation anymore
